+++ testsupport/testcontainers/src/test/kotlin/com/adobe/testing/s3mock/testcontainers/S3MockContainerManualTest.kt
@@ -28,11 +28,11 @@ internal class S3MockContainerManualTest : S3MockContainerTestBase() {
   @BeforeEach
   fun setUp() {
     s3Mock =
-      S3MockContainer(S3MOCK_VERSION).apply {
-        withValidKmsKeys(TEST_ENC_KEYREF)
-        withInitialBuckets(INITIAL_BUCKET_NAMES.joinToString(","))
-        start()
-      }
+      S3MockContainer(S3MOCK_VERSION)
+        .withValidKmsKeys(TEST_ENC_KEYREF)
+        .withInitialBuckets(INITIAL_BUCKET_NAMES.joinToString(","))
+        .withDebugLogging()
+        .apply { start() }
     // Must create S3Client after S3MockContainer is started, otherwise we can't request the random
     // locally mapped port for the endpoint
     s3Client = createS3ClientV2(s3Mock.httpsEndpoint)